requirements.txt
app-version==1.0.1
astroid==2.3.2
botocore==1.13.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.1.7
django-bootstrap-typeahead==1.1.5
django-cart==1.0.4
django-filter==2.2.0
django-filters==0.2.1
django-jsonfield==1.3.1
django-notify-x==0.1.9
-e git+https://github.com/gbozee/django-ravepay.git@cc9fba3440457ee2941bee2db464520eb00ad98e#egg=django_ravepay
django-typeahead==1.0.1
djangorestframework==3.10.3
docopt==0.6.2
docutils==0.15.2
drfdocs==0.0.12b0
future==0.16.0
gcloud==0.17.0
get==2019.4.13
googleapis-common-protos==1.5.8
httplib2==0.12.1
idna==2.7
isort==4.3.20
jmespath==0.9.4
jsonpickle==1.2
jsonschema==2.6.0
jws==0.1.3
lazy-object-proxy==1.4.1
mccabe==0.6.1
num2words==0.5.10
oauth2client==3.0.0
oauthlib==3.0.1
pathspec==0.5.9
Pillow==5.4.1
post==2019.4.13
protobuf==3.6.0
public==2019.4.13
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.4.3
PyJWT==1.7.1
pylint==2.3.1
pyPdf==1.13
PyPDF2==1.26.0
Pyrebase==3.0.27
Pyrebase4==4.3.0
python-dateutil==2.8.0
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
query-string==2019.4.13
request==2019.4.13
requests==2.11.1
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
rsa==3.4.2
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
termcolor==1.1.0
texttable==0.9.1
tinytag==1.2.2
typed-ast==1.4.0
urllib3==1.24.3
validate-email==1.3
wcwidth==0.1.7
websocket-client==0.56.0
wrapt==1.11.1
